Properly get bootstrap working
