+++ src/App.js
@@ -1,31 +1,37 @@
 import React, { Component } from 'react';
-import './App.css';
-import SteamWorkshopItem from './Components/SteamWorkshopItem/SteamWorkshopItem';
+import Button from 'react-bootstrap/Button';
+
+import classes from './App.module.css';
+import SteamWorkshopItem from './components/SteamWorkshop/SteamWorkshopItem/SteamWorkshopItem';
 
 class App extends Component {
   render() {
     return (
 
-      <div className="App">
-      <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"></link>
-      <container>
-        <SteamWorkshopItem
-          clickAction='clickAction'
-          image={'imagex'}
-          title={'title'}
-          summary={'summary'}
-          views={'views'}
-          subscribers={'subscirbers'}
-          favourites={'favourites'}/>
-        <SteamWorkshopItem
-          clickAction='clickAction'
-          image={'imagex'}
-          title={'title'}
-          summary={'summary'}
-          views={'views'}
-          subscribers={'subscirbers'}
-          favourites={'favourites'}/>
-          </container>
+      <div className={classes.App}>
+        <container>
+          <SteamWorkshopItem
+            clickAction='clickAction'
+            image={'imagex'}
+            title={'title'}
+            summary={'summary'}
+            views={'views'}
+            subscribers={'subscirbers'}
+            favourites={'favourites'}
+          />
+          <SteamWorkshopItem
+            clickAction='clickAction'
+            image={'imagex'}
+            title={'title'}
+            summary={'summary'}
+            views={'views'}
+            subscribers={'subscirbers'}
+            favourites={'favourites'}
+          />
+        </container>
+        <Button variant="primary" size="lg" block>
+          Block level button
+        </Button>
       </div>
 
     );
