+++ src/hoc/Aux.js
@@ -1,0 +1,3 @@
+const aux = (props) => props.children;
+
+export default aux;
